add onRenameTable function to dwc
