+++ packages/client/package.json
@@ -24,7 +24,7 @@
 	},
 	"dependencies": {
 		"@data-wrangling-components/core": "^1.2.0",
-		"@data-wrangling-components/react": "^1.4.0",
+		"@data-wrangling-components/react": "^1.4.1",
 		"@data-wrangling-components/utilities": "^1.2.1",
 		"@essex-js-toolkit/boolean-expression-component": "^0.1.1",
 		"@essex-js-toolkit/hooks": "^3.0.2",
